+++ test/unit/metrics.test.ts
@@ -28,6 +28,9 @@
 const Test = require('tapes')(require('tape'))
 import { Histogram } from "prom-client"
 import { Metrics, metricOptionsType } from "../../src/metrics"
+import { Server } from '@hapi/hapi'
+import { Socket } from 'net'
+import Http from 'http'
 
 Test('Metrics Class Test', (metricsTest: any) => {
 
@@ -98,7 +101,7 @@ Test('Metrics Class Test', (metricsTest: any) => {
                     prefix: 'prefixIsInitiated1_',
                     timeout: 1000
                 }
-                
+
                 // metrics.setup(options) <-- we skip this
 
                 let result = metrics.isInitiated()
@@ -207,6 +210,22 @@ Test('Metrics Class Test', (metricsTest: any) => {
         getOptionsTest.end()
     })
 
+    metricsTest.test('getClient should', (getOptionsTest: any) => {
+        getOptionsTest.test('return the client', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                const result = metrics.getClient()
+                test.ok(result, 'Return client')
+                test.end()
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+
+        getOptionsTest.end()
+    })
+
     metricsTest.test('getHistogram should', (getHistogramTest: any) => {
         getHistogramTest.test('return the histogram', async (test: any) => {
             try {
@@ -223,7 +242,7 @@ Test('Metrics Class Test', (metricsTest: any) => {
                     aggregator: 'sum',
                     upperBounds: buckets,
                     buckets,
-                    bucketValues: { '0.010': 0, '0.050': 0, '0.1': 0, '0.5': 0, '1':0 , '2':0 , '3':0 },
+                    bucketValues: { '0.010': 0, '0.050': 0, '0.1': 0, '0.5': 0, '1': 0, '2': 0, '3': 0 },
                     sum: 0,
                     count: 0,
                     hashMap: {},
@@ -233,7 +252,7 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 metrics.setup(options)
                 const result: Histogram<string> = metrics.getHistogram('test_request',
                     histogramConfig.help,
-                    histogramConfig.labelNames, 
+                    histogramConfig.labelNames,
                     histogramConfig.buckets)
                 test.equal(Object.getPrototypeOf(result).constructor.name, histogramConfig.constructor, 'Histogram object is not valid')
                 test.end()
@@ -268,7 +287,7 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 metrics.setup(options)
                 const result: object = metrics.getHistogram('test_request',
                     histogramConfig.help,
-                    histogramConfig.labelNames, 
+                    histogramConfig.labelNames,
                     histogramConfig.buckets)
                 test.fail('Expected an error to be thrown with help param being empty or null')
                 test.end()
@@ -337,14 +356,14 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 metrics.setup(options)
                 const firstResult: object = metrics.getHistogram('test_request',
                     histogramConfig.help,
-                    histogramConfig.labelNames, 
+                    histogramConfig.labelNames,
                     histogramConfig.buckets)
                 const secondResult: object = metrics.getHistogram('test_request',
                     histogramConfig.help,
-                    histogramConfig.labelNames, 
+                    histogramConfig.labelNames,
                     histogramConfig.buckets)
                 test.equal(Object.getPrototypeOf(firstResult).constructor.name, histogramConfig.constructor, 'Histogram object is not valid')
-                test.equal(Object.getPrototypeOf(secondResult).constructor.name, histogramConfig.constructor, 'Histogram object is not valid')   
+                test.equal(Object.getPrototypeOf(secondResult).constructor.name, histogramConfig.constructor, 'Histogram object is not valid')
                 test.end()
             } catch (e) {
                 test.fail(`Error Thrown - ${e}`)
@@ -393,14 +412,14 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 const summaryConfig = {
                     constructor: 'Summary',
                     maxAgeSeconds: 300,
-                    ageBuckets: 2, 
-                    name: `${options.prefix}${metricName}`, 
-                    help: 'Summary for http operation', 
-                    aggregator: 'sum', 
-                    percentiles: [ 0.01, 0.05, 0.1, 0.5, 1, 2, 3 ], 
-                    hashMap: {}, 
-                    labelNames: [ 'success', 'fsp', 'operation', 'source', 'destination' ], 
-                    compressCount: 1000 
+                    ageBuckets: 2,
+                    name: `${options.prefix}${metricName}`,
+                    help: 'Summary for http operation',
+                    aggregator: 'sum',
+                    percentiles: [0.01, 0.05, 0.1, 0.5, 1, 2, 3],
+                    hashMap: {},
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination'],
+                    compressCount: 1000
                 }
 
                 metrics.setup(options)
@@ -432,14 +451,14 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 const summaryConfig = {
                     constructor: 'Summary',
                     maxAgeSeconds: 600,
-                    ageBuckets: 5, 
-                    name: `${options.prefix}${metricName}`, 
-                    help: '', 
-                    aggregator: 'sum', 
-                    percentiles: [ 0.01, 0.05, 0.1, 0.5, 1, 2, 3 ], 
-                    hashMap: {}, 
-                    labelNames: [ 'success', 'fsp', 'operation', 'source', 'destination' ], 
-                    compressCount: 1000 
+                    ageBuckets: 5,
+                    name: `${options.prefix}${metricName}`,
+                    help: '',
+                    aggregator: 'sum',
+                    percentiles: [0.01, 0.05, 0.1, 0.5, 1, 2, 3],
+                    hashMap: {},
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination'],
+                    compressCount: 1000
                 }
 
                 metrics.setup(options)
@@ -470,14 +489,14 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 const summaryConfig = {
                     constructor: 'Summary',
                     maxAgeSeconds: 600,
-                    ageBuckets: 5, 
-                    name: `${options.prefix}${metricName}`, 
-                    help: '', 
-                    aggregator: 'sum', 
-                    percentiles: [ 0.01, 0.05, 0.1, 0.5, 1, 2, 3 ], 
-                    hashMap: {}, 
-                    labelNames: [ 'success', 'fsp', 'operation', 'source', 'destination' ], 
-                    compressCount: 1000 
+                    ageBuckets: 5,
+                    name: `${options.prefix}${metricName}`,
+                    help: '',
+                    aggregator: 'sum',
+                    percentiles: [0.01, 0.05, 0.1, 0.5, 1, 2, 3],
+                    hashMap: {},
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination'],
+                    compressCount: 1000
                 }
 
                 metrics.setup(options)
@@ -505,14 +524,14 @@ Test('Metrics Class Test', (metricsTest: any) => {
                 const summaryConfig = {
                     constructor: 'Summary',
                     maxAgeSeconds: 300,
-                    ageBuckets: 2, 
-                    name: `${options.prefix}${metricName}`, 
-                    help: 'Summary for http operation', 
-                    aggregator: 'sum', 
-                    percentiles: [ 0.01, 0.05, 0.1, 0.5, 1, 2, 3 ], 
-                    hashMap: {}, 
-                    labelNames: [ 'success', 'fsp', 'operation', 'source', 'destination' ], 
-                    compressCount: 1000 
+                    ageBuckets: 2,
+                    name: `${options.prefix}${metricName}`,
+                    help: 'Summary for http operation',
+                    aggregator: 'sum',
+                    percentiles: [0.01, 0.05, 0.1, 0.5, 1, 2, 3],
+                    hashMap: {},
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination'],
+                    compressCount: 1000
                 }
 
                 metrics.setup(options)
@@ -545,6 +564,324 @@ Test('Metrics Class Test', (metricsTest: any) => {
         })
 
         getSummaryTest.end()
+    })
+
+    metricsTest.test('getCounter should', (getCounterTest: any) => {
+        getCounterTest.test('return the counter', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                const options: metricOptionsType = {
+                    prefix: 'prefixCounter1_',
+                    timeout: 1000
+                }
+
+                const counterConfig = {
+                    constructor: 'Counter',
+                    name: 'prefixCounter1_test_counter',
+                    help: 'Counter for http operation',
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination']
+                }
+
+                metrics.setup(options)
+                const result = metrics.getCounter(
+                    'test_counter',
+                    counterConfig.help,
+                    counterConfig.labelNames
+                )
+                test.equal(Object.getPrototypeOf(result).constructor.name, counterConfig.constructor, 'Counter object is not valid')
+                test.end()
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+
+        getCounterTest.test('return the counter if help param is empty string', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                const options: metricOptionsType = {
+                    prefix: 'prefixCounter2_',
+                    timeout: 1000
+                }
+
+                const counterConfig = {
+                    constructor: 'Counter',
+                    name: 'prefixCounter2_test_counter',
+                    help: '',
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination']
+                }
+
+                metrics.setup(options)
+                metrics.getCounter(
+                    'test_counter',
+                    counterConfig.help,
+                    counterConfig.labelNames
+                )
+
+                test.fail('Expected an error to be thrown with help param being empty or null')
+                test.end()
+            } catch (e) {
+                test.equal(e.message, 'Couldn\'t get counter for test_counter')
+                test.end()
+            }
+        })
+
+        getCounterTest.test('return the counter if help param is null', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                const options: metricOptionsType = {
+                    prefix: 'prefixCounter3_',
+                    timeout: 1000
+                }
+
+                metrics.setup(options)
+                metrics.getCounter(
+                    'test_counter'
+                )
+
+                test.fail('Expected an error to be thrown with help param being empty or null')
+                test.end()
+            } catch (e) {
+                test.equal(e.message, 'Couldn\'t get counter for test_counter')
+                test.end()
+            }
+        })
+
+        getCounterTest.test('return the existing counter', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                const options: metricOptionsType = {
+                    prefix: 'prefixCounter4_',
+                    timeout: 1000
+                }
+
+                const counterConfig = {
+                    constructor: 'Counter',
+                    name: 'prefixCounter4_test_counter',
+                    help: 'Counter for http operation',
+                    labelNames: ['success', 'fsp', 'operation', 'source', 'destination']
+                }
+
+                metrics.setup(options)
+                const firstResult = metrics.getCounter(
+                    'test_counter',
+                    counterConfig.help,
+                    counterConfig.labelNames
+                )
+                const secondResult = metrics.getCounter(
+                    'test_counter',
+                    counterConfig.help,
+                    counterConfig.labelNames
+                )
+                test.equal(Object.getPrototypeOf(firstResult).constructor.name, counterConfig.constructor, 'Counter object is not valid')
+                test.equal(Object.getPrototypeOf(secondResult).constructor.name, counterConfig.constructor, 'Counter object is not valid')
+                test.end()
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+        getCounterTest.end()
+    })
+
+    metricsTest.test('metrics plugin should', (pluginTest: any) => {
+        pluginTest.test('register http metrics', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                metrics.getDefaultRegister().clear()
+                const options: metricOptionsType = {
+                    prefix: 'plugin1_',
+                    timeout: 1000
+                }
+                metrics.setup(options)
+                const server = new Server({ port: 0 })
+                await server.register(metrics.plugin)
+                server.route({
+                    method: 'GET',
+                    path: '/test',
+                    handler: () => 'test'
+                })
+                server.route({
+                    method: 'GET',
+                    path: '/{p*}',
+                    handler: () => 'test'
+                })
+                await server.start()
+                test.ok(server, 'Server is started')
+                await server.inject({ method: 'GET', url: '/test' })
+                await server.inject({ method: 'GET', url: '/coverage' })
+                await server.inject({ method: 'GET', url: '/live' })
+                const metricsResponse = await server.inject({
+                    method: 'GET',
+                    url: '/metrics'
+                })
+                test.equal(metricsResponse.statusCode, 200, 'Metrics status code is 200')
+                test.ok(metricsResponse.payload.match(/http_requests_total/), 'Total number of http requests metrics is returned')
+                test.ok(metricsResponse.payload.match(/http_request_duration_seconds/), 'Duration of http requests metric is returned')
+                test.ok(metricsResponse.payload.match(/http_request_duration_histogram_seconds/), 'Duration of http requests histogram is returned')
+                test.ok(metricsResponse.payload.match(/http_requests_current/), 'Number of requests currently running metric is returned')
+                test.ok(metricsResponse.payload.match(/http_connections_current/), 'Number of connections currently established metric is returned')
+                await server.stop()
+                test.end()
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+        pluginTest.test('limit max pending requests', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                metrics.getDefaultRegister().clear()
+                const options: metricOptionsType = {
+                    prefix: 'plugin2_',
+                    timeout: 1000,
+                    maxRequestsPending: 1
+                }
+                metrics.setup(options)
+                const server = new Server({ port: 0 })
+                await server.register(metrics.plugin)
+                server.route({
+                    method: 'GET',
+                    path: '/health',
+                    handler: () => 'ready'
+                })
+                server.route({
+                    method: 'GET',
+                    path: '/delay',
+                    handler: async () => {
+                        await new Promise(resolve => setTimeout(resolve, 2000))
+                        return 'ready'
+                    }
+                })
+                await server.start()
+                test.ok(server, 'Server is started')
+
+                const readyResponse = await server.inject({
+                    method: 'GET',
+                    url: '/health'
+                })
+                test.equal(readyResponse.statusCode, 200, 'Ready status code is 200')
+
+                const delayResponse = server.inject({
+                    method: 'GET',
+                    url: '/delay'
+                })
+
+                await new Promise(resolve => setTimeout(resolve, 1000)) // wait for the request to be processed
+                const notReady = await server.inject({
+                    method: 'GET',
+                    url: '/health'
+                })
+                test.equal(notReady.statusCode, 503, 'Requests limit reached')
+                await delayResponse;
+                await server.stop()
+                test.end()
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+        pluginTest.test('limit connections', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                metrics.getDefaultRegister().clear()
+                const options: metricOptionsType = {
+                    prefix: 'plugin3_',
+                    timeout: 1000,
+                    maxConnections: 1
+                }
+                metrics.setup(options)
+                const server = new Server({ port: 0 })
+                await server.register(metrics.plugin)
+                server.route({
+                    method: 'GET',
+                    path: '/health',
+                    handler: () => 'ready'
+                })
+                await server.start()
+                test.ok(server, 'Server is started')
+                const socket = new Socket()
+                socket.on('connect', async () => {
+                    test.pass('Connection established')
+                    try {
+                        const readyResponse = await server.inject({
+                            method: 'GET',
+                            url: '/health'
+                        })
+                        test.equal(readyResponse.statusCode, 503, 'Connection limit reached')
+                        socket.destroy()
+                    }
+                    catch (e) {
+                        test.fail(`Error Thrown - ${e}`)
+                        test.end()
+                    }
+                })
+                socket.on('close', async () => {
+                    await new Promise(resolve => setTimeout(resolve, 1000)) // wait for the connection to close
+                    test.pass('Connection closed')
+                    try {
+                        const readyResponse = await server.inject({
+                            method: 'GET',
+                            url: '/health'
+                        })
+                        test.equal(readyResponse.statusCode, 200, 'Ready status code is 200')
+                        await server.stop()
+                        test.end()
+                    } catch (e) {
+                        test.fail(`Error Thrown - ${e}`)
+                        test.end()
+                    }
+                });
+                socket.connect(Number(server.info.port), 'localhost');
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+        pluginTest.test('handle aborted requests', async (test: any) => {
+            try {
+                const metrics: Metrics = new Metrics()
+                metrics.getDefaultRegister().clear()
+                const options: metricOptionsType = {
+                    prefix: 'plugin4_',
+                    timeout: 1000
+                }
+                metrics.setup(options)
+                const server = new Server({ port: 0 })
+                await server.register(metrics.plugin)
+                let req: Http.ClientRequest;
+                server.route({
+                    method: 'GET',
+                    path: '/',
+                    handler: (request) => new Promise((resolve) => {
+                        req.destroy();
+                        request.events.once('disconnect', () => {
+                            resolve('ok')
+                        });
+                    })
+                });
+
+                await server.start()
+                test.ok(server, 'Server is started')
+
+                req = Http.request({
+                    hostname: 'localhost',
+                    port: server.info.port,
+                    method: 'get'
+                });
+                req.on('error', () => { });
+                req.end();
+
+                const [request] = await server.events.once('response');
+                test.ok(request.response.isBoom, 'Request is aborted');
+                await server.stop()
+                test.end();
+            } catch (e) {
+                test.fail(`Error Thrown - ${e}`)
+                test.end()
+            }
+        })
+        pluginTest.end()
     })
 
     metricsTest.end()